requirements.txt
streamlit
langchain
langchain-community
PyPDF2
faiss-cpu
ollama
python-dotenv